Adding a missing definition for joypad code. Whoops
Also adding a new function that allows code to pass in a date and get whether the current date is past or not this date. This is useful for limiting usage of a feature for things like demo's, or time released dlc.

NOTE - this is ONLY populated for Android and Windows currently. I'll fix the IOS version shortly, but for WebOS and RIM, someone else will have to populate that.

git-svn-id: svn://rtsoft.com/rtsvn@965 353e56fe-9613-0410-8469-b96ad8e6f29c